Add basic Docker setup for Laravel application
A Docker environment has been set up for the Laravel application. This includes configuration files for MySQL, Nginx, and PHP-FPM, along with a docker-compose file to manage the services. In addition, there's a Makefile to simplify Docker commands and environment variables have been defined for Docker in a .env.example file.